Feature: Custom Styling for Embedded Downloadable Zip Files (#50)
Adds unique appearance to embedded zip files, styling these like a download button.

`.zip` and `.7z` files will be the main filetype (and at this stage, only filetype) used for downloadable links for tutorials and the like. This makes it more obvious to users that these are meant to be clicked and downloaded.