Merge `SEO-Optimazation`  branch into `Main` (#58)
* Add `file-naming-guidelines.md`

Added `file-naming-guidelines.md` to act as a reference while restructuring the file-names. This is in order to better help search engines like Google index the wiki.

* Update: All `.md` files updated

Updated all `.md` filenames to be lowercase, with spaces replaced by hyphens.

* Update: rename assets for patch & dialogue tutorials

Renamed assets for:
- `guides/patches/esp-replacer-patch-beginner-tutorial.md`
- `guides/quest-design/branching-dialogue-tutorial.md`
Asset names made lowercase, spaces replaced with hyphens, and given names are descriptive.

* Update: `description` frontmatter added to hompage

Root `index.md` file lacked a `description:` frontmatter field. This field determines the description when a link to the homepage is shared on social media or displayed in search results.

* Chore: Old github social preview image deleted

This has been replaced with `quartz/static/og-image.png`

* Update: reorganisee assets directories and update some articles

Updated assets for the following articles:
- tes3cs-getting-started: `dungeon-tutorial`
- tes3cs-getting-started: `landscape-tutorial`
- morrowind-interiors-project

Rearranged assets directory so it correctly follows structure of the rest of the wiki

* Update: rename mlox docs images

* Update: assets for `how-to-contribute.md` updated

- renamed filenames to descriptive names, lowercase, with spaces replaced with hyphens
- deleted unused images (checked backlinks to ensure they were not embedded in any pages)

* Update: `simple-texture-animation-in-blender.md` assets renames

- renamed assets in `simple-texture-animation-in-blender.md` to optimize for SEO
- updated link to 'MMW Design & Planning' to 'Contributing' on home page.