Fix: Changed selector from `~=` to `$=`
Changed `[data-embed-alias` selector from `~=clean]` to `$=clean]`. This way, it only selects "clean" if it is the last phrase in the embed alt-text.

This is because every note in the wiki has an `alias:` defined in frontmatter. When entering the note title in Obsidian, it will make e.g., `![[cool-note-name#^adfe39]]` into `![[cool-note-name#^adfe39|Cool Note Name]]`. If a note contained the word "clean" in its alias, Quartz would consider this the keyword to make the transclusion `clean` even if that was not the author's intention.

For embedded/transcluded notes and blocks, the alias is not actually relevant as the title will not be displayed; therefore, it can be removed. However, because Obsidian will populate it anyway for embeds, if an author does not manually remove the `|alias]]` from the embed wikilink, they may encounter issues. By changing the selector to `$=`, only the last word of the alias will be considered for styling the embed, so the `clean` property can be added to the end of the embed title without needing to remove the alias name.

- E.g., `![[cool-note-name#^adfe39|Cool Note Name clean]]` gets rendered as a clean transclusion as the last word is "clean"
- `![[cool-note-name#^adfe39|clean Cool Note Name]]` does not get rendered as a transclusion as the last word is "name"