Fix: add `left` and `right` float + revert selector to `~=`
Instead of placing embeds in a `<blockquote>` like Quartz, Obsidian places embeds in a `<span>`. When embedding an element which already floats left or right, the embedded element will float accordingly, but its `<span>` container takes up the full page-width, thereby preventing the embedded element from floating around other content on the page. As a workaround, I added `left` and `right` alias metadata for embeds; this is redundant in Quartz, but allows Obsidian to render the embeds in the same way as Quartz - WYSIWYG.

Due to adding the `left` and `right` metadata, the `$=` attribute selectors needed to be reverted to `~=` selectors, as it should not matter what order users place `|clean right`/`|left clean` etc.