Feature: Add option to hide transclusion styling using `|clean` (#61)
Adding the alias `|clean` to a transclusion hides the blockquote styling.

Additionally, added `|left` and `|right` alias to float embeds left or right. This is required if embedding a block (such as an infobox callout) which already has a float property, because in Obsidian the `<span>` containing the embed will otherwise take up 100% page-width and prevent the floating embed from actually wrapping around other content. Quartz is unaffected as it handles transclusions differently, so the `left` and `right` aliases are not defined in the Quartz scss. This is purely a WYSIWYG compromise with Obsidian.

This update also marks a change in approach for creating custom scss for Quartz. Because the `styles/custom.scss` file has bloated in size and is becoming more difficult to navigate, I opted to utilise scss modules. The new Quartz css for transclusions can be found in `styles/custom/transclusions.scss`, which is imported as a module into `custom.scss`.

Closes #47