Fix: Reduce line height for txt-s tables
Reduce line-height of tables to 1.6rem if table is inside a callout with `|txt-s` adjustment. Closes #71

Quartz's default line-height for tables is 2rem. If using the `|txt-s` callout adjustment to reduce font-size of callout content, tables inside the callout do not have their line-height reduced to the scale of the reduced text size, so lines appear abnormally far apart.