.gitignore: allow `.obsidian/types.json`
`.obsidian/types.json` defines types of Obsidian properties. If one creates a custom property (the default property type is 'text') and tells it to behave as a 'list', this updates `types.json`. If this file is ignored, those custom properties will revert to default Property Type so Obsidian will throw an error for the YAML properties, as the property type of 'text' would be formatted like a 'list'.